requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
bbox-visualizer==0.1.0
bbox-visualizer2==0.0.4
certifi==2023.7.22
charset-normalizer==3.2.0
easyocr==1.7.1
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.10.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
imageio==2.31.6
Jinja2==3.1.2
lazy_loader==0.3
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2
ninja==1.11.1.1
numpy==1.26.1
opencv-python-headless==4.8.1.78
packaging==23.2
Pillow==10.0.0
pyclipper==1.3.0.post5
pytesseract==0.3.10
python-bidi==0.4.2
python-telegram-bot==20.6
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.3
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
sympy==1.12
terminal-img==0.0.3
tifffile==2023.9.26
torch==2.1.0
torchvision==0.16.0
typing_extensions==4.8.0
urllib3==2.0.4
yarl==1.9.2
